separated contact into customer and supplier. Furthermore, created controller for contact group, customer and supplier

+++ app/models/contact_group.rb
@@ -22,6 +22,11 @@ class ContactGroup < ActiveRecord::Base
   end
   
   def delete_object
+    if self.contacts.count != 0
+      self.errors.add(:generic_errors, "Sudah ada contact")
+      return self 
+    end
+    
     self.destroy
   end
